Removed all the old FindInFiles functionality, classes, signals, UI. Reorganized new code and started code cleanup and documentation.
