+++ src/ui/Search/frmsearchreplace.cpp
@@ -32,28 +32,12 @@ frmSearchReplace::frmSearchReplace(TopEditorContainer *topEditorContainer, QWidg
     ui->cmbSearch->setCurrentText("");
     ui->cmbReplace->addItems(s.Search.getReplaceHistory());
     ui->cmbReplace->setCurrentText("");
-    ui->cmbLookIn->addItems(s.Search.getFileHistory());
-    ui->cmbLookIn->setCurrentText("");
-    ui->cmbFilter->addItems(s.Search.getFilterHistory());
-    ui->cmbFilter->setCurrentText("");
 
     connect(ui->cmbSearch->lineEdit(), &QLineEdit::textEdited, this, &frmSearchReplace::on_searchStringEdited);
-    connect(ui->cmbSearch->lineEdit(), &QLineEdit::returnPressed, this, [=]() {
-        if (ui->actionFind_in_files->isChecked()) {
-            on_btnFindAll_clicked();
-        } else {
-            on_btnFindNext_clicked();
-        }
-    });
+    connect(ui->cmbSearch->lineEdit(), &QLineEdit::returnPressed, this, &frmSearchReplace::on_btnFindNext_clicked);
     connect(ui->cmbReplace->lineEdit(), &QLineEdit::returnPressed, this, &frmSearchReplace::on_btnFindNext_clicked);
-    connect(ui->cmbLookIn->lineEdit(), &QLineEdit::returnPressed, this, &frmSearchReplace::on_btnFindAll_clicked);
-    connect(ui->cmbFilter->lineEdit(), &QLineEdit::returnPressed, this, &frmSearchReplace::on_btnFindAll_clicked);
-
 
     connect(ui->actionFind_in_files, &QAction::triggered, this, &frmSearchReplace::advancedFindRequested);
-
-
-    ui->cmbFilter->lineEdit()->setPlaceholderText("*.ext1, *.ext2, ...");
 
     ui->actionFind->setIcon(IconProvider::fromTheme("edit-find"));
     ui->actionReplace->setIcon(IconProvider::fromTheme("edit-find-replace"));
@@ -206,143 +190,6 @@ int frmSearchReplace::selectAll(QString string, SearchHelpers::SearchMode search
     return count.toInt();
 }
 
-void frmSearchReplace::sessionCleanup()
-{
-    if (m_session != nullptr) {
-        m_session->threadSearch = nullptr;
-        m_session->threadReplace = nullptr;
-        if (m_session->msgBox != nullptr) {
-            m_session->msgBox->hide();
-            m_session->msgBox->deleteLater();
-            m_session->msgBox = nullptr;
-        }
-        m_session->deleteLater();
-        m_session = nullptr;
-    }
-}
-
-void frmSearchReplace::displayThreadErrorMessageBox(const QString &message, int &operation)
-{
-    operation = QMessageBox::warning(
-                this,
-                tr("Error"),
-                message,
-                QMessageBox::Abort | QMessageBox::Retry | QMessageBox::Ignore,
-                QMessageBox::Retry);
-}
-
-void frmSearchReplace::handleSearchResult(const FileSearchResult::SearchResult &result)
-{
-    sessionCleanup();
-    emit fileSearchResultFinished(result);
-}
-
-void frmSearchReplace::handleReplaceResult(int replaceCount, int fileCount, bool stopped)
-{
-    sessionCleanup();
-    QApplication::processEvents();
-    if (!stopped) {
-        QMessageBox::information(this,
-                                 tr("Replace in files"),
-                                 tr("%1 occurrences replaced in %2 files.").arg(replaceCount).arg(fileCount));
-    } else {
-        QMessageBox::information(this,
-                                 tr("Replace in files"),
-                                 tr("%1 occurrences replaced in %2 files, but the replacement has been canceled before it could finish.").arg(replaceCount).arg(fileCount));
-    }
-}
-
-void frmSearchReplace::handleError(const QString &e)
-{
-    if (m_session->msgBox != nullptr) {
-        m_session->msgBox->setTitle(tr("Error"));
-        m_session->msgBox->setText(e);
-    }
-}
-
-void frmSearchReplace::handleProgress(const QString &file, bool replace)
-{
-    if (m_session->msgBox != nullptr) {
-        if (replace) {
-            m_session->msgBox->setText(tr("Replacing in ").append("%1").arg(file));
-        } else {
-            m_session->msgBox->setText(tr("Searching in ").append("%1").arg(file));
-        }
-    }
-}
-
-void frmSearchReplace::handleReplaceInFiles(const FileSearchResult::SearchResult &result)
-{
-    QString replacement = ui->cmbReplace->currentText();
-    m_session->msgBox->hide();
-    m_session->msgBox->deleteLater();
-    m_session->msgBox = new dlgSearching(this);
-    m_session->msgBox->setTitle(tr("Replacing..."));
-    m_session->threadReplace = new ReplaceInFilesWorker(this, result, replacement);
-
-    connect(m_session->threadReplace, SIGNAL(finished()), m_session->threadReplace, SLOT(deleteLater()));
-    connect(m_session->threadReplace, &ReplaceInFilesWorker::error, this, &frmSearchReplace::handleError);
-    connect(m_session->threadReplace, &ReplaceInFilesWorker::errorReadingFile, this, &frmSearchReplace::displayThreadErrorMessageBox, Qt::BlockingQueuedConnection);
-    connect(m_session->threadReplace, &ReplaceInFilesWorker::progress, this, &frmSearchReplace::handleProgress);
-    connect(m_session->threadReplace, &ReplaceInFilesWorker::resultReady, this, &frmSearchReplace::handleReplaceResult);
-    connect(this, &frmSearchReplace::stopReplaceInFiles, m_session->threadReplace, &ReplaceInFilesWorker::stop, Qt::DirectConnection);
-    m_session->threadReplace->start();
-    if (m_session->msgBox->exec()) {
-        emit stopReplaceInFiles();
-    }
-    sessionCleanup();
-}
-
-bool frmSearchReplace::confirmReplaceInFiles(const QString &path, const QStringList &filters)
-{
-    return (QMessageBox::warning(this,
-                        tr("Replace in files"),
-                        tr("Are you sure you want to replace all occurrences in %1 for file types %2?")
-                           .arg(path)
-                           .arg(filters.isEmpty() ? "*" : filters.join(", ")),
-                        QMessageBox::Ok | QMessageBox::Cancel,
-                        QMessageBox::Cancel)
-                  == QMessageBox::Ok);
-}
-
-void frmSearchReplace::searchReplaceInFiles(const QString &string, const QString &path, const QStringList &filters, const SearchHelpers::SearchMode &searchMode, const SearchHelpers::SearchOptions &searchOptions, bool replaceMode)
-{
-    if (m_session == nullptr) {
-        m_session = new SearchInFilesSession(this);
-    }
-    if (!string.isEmpty()) {
-        if (replaceMode && !confirmReplaceInFiles(path, filters)) {
-            sessionCleanup();
-            return;
-        }
-        m_session->threadSearch = new SearchInFilesWorker(this, string, path, filters, searchMode, searchOptions);
-        m_session->msgBox = new dlgSearching(this);
-        m_session->msgBox->setTitle(tr("Searching..."));
-        m_session->msgBox->setWindowTitle(m_session->msgBox->title());
-
-        connect(m_session->threadSearch, SIGNAL(finished()), m_session->threadSearch, SLOT(deleteLater()));
-        connect(m_session->threadSearch, &SearchInFilesWorker::error, this, &frmSearchReplace::handleError);
-        connect(m_session->threadSearch, &SearchInFilesWorker::errorReadingFile, this, &frmSearchReplace::displayThreadErrorMessageBox, Qt::BlockingQueuedConnection);
-        connect(m_session->threadSearch, &SearchInFilesWorker::progress, this, &frmSearchReplace::handleProgress);
-        connect(this, &frmSearchReplace::stopSearchInFiles, m_session->threadSearch, &SearchInFilesWorker::stop, Qt::DirectConnection);
-        //Send results to a different location in the event of replaceMode.
-        if (replaceMode) {
-            connect(m_session->threadSearch, &SearchInFilesWorker::resultReady, this, &frmSearchReplace::handleReplaceInFiles);
-        } else {
-            connect(m_session->threadSearch, &SearchInFilesWorker::resultReady, this, &frmSearchReplace::handleSearchResult);
-        }
-        m_session->threadSearch->start();
-
-        int cancel = m_session->msgBox->exec();
-        if (cancel) {
-            emit stopSearchInFiles();
-        }
-        if (!replaceMode || cancel) {
-            sessionCleanup();
-        }
-    }
-}
-
 SearchHelpers::SearchMode frmSearchReplace::searchModeFromUI()
 {
     if (ui->radSearchPlainText->isChecked())
@@ -366,8 +213,6 @@ SearchHelpers::SearchOptions frmSearchReplace::searchOptionsFromUI()
         searchOptions.MatchCase = true;
     if (ui->chkMatchWholeWord->isChecked())
         searchOptions.MatchWholeWord = true;
-    if (ui->chkIncludeSubdirs->isChecked())
-        searchOptions.IncludeSubDirs = true;
 
     return searchOptions;
 }
@@ -470,28 +315,6 @@ void frmSearchReplace::on_actionFind_toggled(bool /*on*/)
     manualSizeAdjust();
 }
 
-void frmSearchReplace::on_actionFind_in_files_toggled(bool on)
-{
-    ui->lblReplace->setVisible(on);
-    ui->cmbReplace->setVisible(on);
-    ui->lblLookIn->setVisible(on);
-    ui->cmbLookIn->setVisible(on);
-    ui->lblFilter->setVisible(on);
-    ui->cmbFilter->setVisible(on);
-    ui->btnLookInBrowse->setVisible(on);
-    ui->btnFindAll->setVisible(on);
-    ui->lblSpacer1->setVisible(on);
-    ui->btnReplaceAllInFiles->setVisible(on);
-    ui->chkIncludeSubdirs->setVisible(on);
-    ui->btnFindNext->setVisible(!on);
-    ui->btnFindPrev->setVisible(!on);
-    ui->btnSelectAll->setVisible(!on);
-
-    ui->cmbSearch->setFocus();
-
-    manualSizeAdjust();
-}
-
 void frmSearchReplace::manualSizeAdjust()
 {
     int curX = geometry().x();
@@ -563,61 +386,6 @@ void frmSearchReplace::on_searchStringEdited(const QString &/*text*/)
 
     // Workaround. See comment in setSearchText().
     ui->cmbSearch->setAutoCompletion(true);
-}
-
-void frmSearchReplace::on_btnFindAll_clicked()
-{
-    searchReplaceInFiles(ui->cmbSearch->currentText(),
-                  ui->cmbLookIn->currentText(),
-                  fileFiltersFromUI(),
-                  searchModeFromUI(),
-                  searchOptionsFromUI());
-
-    addToSearchHistory(ui->cmbSearch->currentText());
-    addToFileHistory(ui->cmbLookIn->currentText());
-    addToFilterHistory(ui->cmbFilter->currentText());
-}
-
-void frmSearchReplace::on_btnLookInBrowse_clicked()
-{
-    QString defaultDir = ui->cmbLookIn->currentText();
-    if (defaultDir.isEmpty()) {
-        QFileInfo file(currentEditor()->fileName().toLocalFile());
-        defaultDir = file.absolutePath();
-    }
-
-    QString dir = QFileDialog::getExistingDirectory(this, tr("Look in"),
-                                                     defaultDir,
-                                                     QFileDialog::ShowDirsOnly
-                                                     | QFileDialog::DontResolveSymlinks);
-
-    if (!dir.isEmpty()) {
-        ui->cmbLookIn->setCurrentText(dir);
-    }
-}
-
-void frmSearchReplace::on_btnReplaceAllInFiles_clicked()
-{
-    searchReplaceInFiles(ui->cmbSearch->currentText(),
-                   ui->cmbLookIn->currentText(),
-                   fileFiltersFromUI(),
-                   searchModeFromUI(),
-                   searchOptionsFromUI(),
-                   true);
-
-    addToSearchHistory(ui->cmbSearch->currentText());
-    addToReplaceHistory(ui->cmbReplace->currentText());
-    addToFileHistory(ui->cmbLookIn->currentText());
-    addToFilterHistory(ui->cmbFilter->currentText());
-}
-
-QStringList frmSearchReplace::fileFiltersFromUI()
-{
-    QStringList filters = ui->cmbFilter->currentText().split(",", QString::SkipEmptyParts);
-    for (int i = 0; i < filters.count(); i++) {
-        filters[i] = filters[i].trimmed();
-    }
-    return filters;
 }
 
 /**
