+++ src/ui/Search/searchstring.cpp
@@ -3,13 +3,12 @@
 
 QString SearchString::toRaw(const QString &data, const SearchHelpers::SearchMode &searchMode, const SearchHelpers::SearchOptions &searchOptions)
 {
-    QString rawSearch = data;
+    QString rawSearch = toRegex(data, searchOptions.MatchWholeWord);
+
     if (searchMode == SearchHelpers::SearchMode::SpecialChars) {
-        rawSearch = toRegex(data, searchOptions.MatchWholeWord);
         rawSearch = rawSearch.replace("\\\\", "\\");
-    } else if (searchMode == SearchHelpers::SearchMode::PlainText){
-        rawSearch = toRegex(data, searchOptions.MatchWholeWord);
     }
+
     return rawSearch;
 }
