+++ src/ui/include/mainwindow.h
@@ -9,7 +9,6 @@
 #include "include/Search/frmsearchreplace.h"
 #include <functional>
 #include "QtPrintSupport/QPrinter"
-#include "include/Search/filesearchresultswidget.h"
 #include "include/Extensions/extension.h"
 #include "include/nqqsettings.h"
 #include "include/Search/advancedsearchdock.h"
@@ -161,8 +160,6 @@ private slots:
     void on_actionDuplicate_Line_triggered();
     void on_actionMove_Line_Up_triggered();
     void on_actionMove_Line_Down_triggered();
-    void on_fileSearchResultFinished(FileSearchResult::SearchResult result);
-    void on_resultMatchClicked(const QString &fileName, int startLine, int startCol, int endLine, int endCol);
     void on_actionTrim_Trailing_Space_triggered();
     void on_actionTrim_Leading_Space_triggered();
     void on_actionTrim_Leading_and_Trailing_Space_triggered();
@@ -199,7 +196,6 @@ private:
     NqqSettings&          m_settings;
     frmSearchReplace*     m_frmSearchReplace = 0;
     bool                  m_overwrite = false; // Overwrite mode vs Insert mode
-    FileSearchResultsWidget* m_fileSearchResultsWidget;
     QString               m_workingDirectory;
     QMap<QSharedPointer<Extensions::Extension>, QMenu*> m_extensionMenus;
