+++ src/ui/ui.pro
@@ -80,16 +80,9 @@ SOURCES += main.cpp\
     frmindentationmode.cpp \
     singleapplication.cpp \
     localcommunication.cpp \
-    Search/filesearchresultswidget.cpp \
     Search/frmsearchreplace.cpp \
-    Search/searchinfilesworker.cpp \
     Search/searchstring.cpp \
-    Search/replaceinfilesworker.cpp \
-    Search/replaceworker.cpp \
-    Search/searchworker.cpp \
     Search/advancedsearchdock.cpp \
-    Search/dlgsearching.cpp \
-    Search/searchresultsitemdelegate.cpp \
     Extensions/extension.cpp \
     frmlinenumberchooser.cpp \
     Extensions/extensionsserver.cpp \
@@ -106,7 +99,10 @@ SOURCES += main.cpp\
     Sessions/sessions.cpp \
     Sessions/persistentcache.cpp \
     nqqsettings.cpp \  
-    nqqrun.cpp
+    nqqrun.cpp \
+    Search/filesearcher.cpp \
+    Search/filereplacer.cpp \
+    Search/searchobjects.cpp
 
 HEADERS  += include/mainwindow.h \
     include/topeditorcontainer.h \
@@ -127,18 +123,10 @@ HEADERS  += include/mainwindow.h \
     include/frmindentationmode.h \
     include/singleapplication.h \
     include/localcommunication.h \
-    include/Search/filesearchresultswidget.h \
-    include/Search/filesearchresult.h \
     include/Search/frmsearchreplace.h \
-    include/Search/searchinfilesworker.h \
-    include/Search/replaceinfilesworker.h \
-    include/Search/replaceworker.h \
-    include/Search/searchworker.h \
     include/Search/advancedsearchdock.h \
     include/Search/searchhelpers.h \
     include/Search/searchstring.h \
-    include/Search/dlgsearching.h \
-    include/Search/searchresultsitemdelegate.h \
     include/Extensions/extension.h \
     include/frmlinenumberchooser.h \
     include/Extensions/extensionsserver.h \
@@ -155,7 +143,10 @@ HEADERS  += include/mainwindow.h \
     include/Sessions/sessions.h \
     include/Sessions/persistentcache.h \
     include/nqqsettings.h \
-    include/nqqrun.h
+    include/nqqrun.h \
+    include/Search/filesearcher.h \
+    include/Search/searchobjects.h \
+    include/Search/filereplacer.h
 
 FORMS    += mainwindow.ui \
     frmabout.ui \
